Refactor form condition editor client code
