+++ phpunit/functional/Glpi/Form/Export/FormSerializerTest.php
@@ -39,12 +39,12 @@ use Computer;
 use Entity;
 use Glpi\Form\Category;
 use Glpi\Form\Comment;
-use Glpi\Form\ConditionalVisiblity\ConditionData;
-use Glpi\Form\ConditionalVisiblity\FormData;
-use Glpi\Form\ConditionalVisiblity\LogicOperator;
-use Glpi\Form\ConditionalVisiblity\Type;
-use Glpi\Form\ConditionalVisiblity\ValueOperator;
-use Glpi\Form\ConditionalVisiblity\VisibilityStrategy;
+use Glpi\Form\Condition\ConditionData;
+use Glpi\Form\Condition\FormData;
+use Glpi\Form\Condition\LogicOperator;
+use Glpi\Form\Condition\Type;
+use Glpi\Form\Condition\ValueOperator;
+use Glpi\Form\Condition\VisibilityStrategy;
 use Glpi\Form\Destination\CommonITILField\AssociatedItemsField;
 use Glpi\Form\Destination\CommonITILField\AssociatedItemsFieldConfig;
 use Glpi\Form\Destination\CommonITILField\AssociatedItemsFieldStrategy;
