+++ src/Glpi/Controller/Form/RendererController.php
@@ -41,8 +41,8 @@ use Glpi\Exception\Http\BadRequestHttpException;
 use Glpi\Exception\Http\NotFoundHttpException;
 use Glpi\Form\AccessControl\FormAccessControlManager;
 use Glpi\Form\AccessControl\FormAccessParameters;
-use Glpi\Form\ConditionalVisiblity\Engine;
-use Glpi\Form\ConditionalVisiblity\EngineInput;
+use Glpi\Form\Condition\Engine;
+use Glpi\Form\Condition\EngineInput;
 use Glpi\Form\Form;
 use Glpi\Form\ServiceCatalog\ServiceCatalog;
 use Glpi\Http\Firewall;
