+++ src/Glpi/Form/Condition/EngineOutput.php
@@ -32,7 +32,7 @@
  * ---------------------------------------------------------------------
  */
 
-namespace Glpi\Form\ConditionalVisiblity;
+namespace Glpi\Form\Condition;
 
 use JsonSerializable;
 use Override;
