+++ src/Glpi/Form/Condition/LogicOperator.php
@@ -32,7 +32,7 @@
  * ---------------------------------------------------------------------
  */
 
-namespace Glpi\Form\ConditionalVisiblity;
+namespace Glpi\Form\Condition;
 
 enum LogicOperator: string
 {
