+++ src/Glpi/Form/Condition/NumberConditionTrait.php
@@ -32,11 +32,11 @@
  * ---------------------------------------------------------------------
  */
 
-namespace Glpi\Form\ConditionalVisiblity;
+namespace Glpi\Form\Condition;
 
 use Override;
 
-trait StringConditionTrait
+trait NumberConditionTrait
 {
     #[Override]
     public function getSupportedValueOperators(): array
@@ -44,15 +44,17 @@ trait StringConditionTrait
         return [
             ValueOperator::EQUALS,
             ValueOperator::NOT_EQUALS,
-            ValueOperator::CONTAINS,
-            ValueOperator::NOT_CONTAINS,
+            ValueOperator::GREATER_THAN,
+            ValueOperator::GREATER_THAN_OR_EQUALS,
+            ValueOperator::LESS_THAN,
+            ValueOperator::LESS_THAN_OR_EQUALS,
         ];
     }
 
     #[Override]
     public function getInputTemplateKey(): InputTemplateKey
     {
-        return InputTemplateKey::STRING;
+        return InputTemplateKey::NUMBER;
     }
 
     #[Override]
@@ -61,19 +63,20 @@ trait StringConditionTrait
         ValueOperator $operator,
         mixed $b,
     ): bool {
-        // Normalize strings.
-        $a = strtolower(strval($a));
-        $b = strtolower(strval($b));
+        // Normalize values.
+        $a = (float) $a;
+        $b = (float) $b;
 
         return match ($operator) {
-            ValueOperator::EQUALS       => $a === $b,
-            ValueOperator::NOT_EQUALS   => $a !== $b,
-            ValueOperator::CONTAINS     => str_contains($b, $a),
-            ValueOperator::NOT_CONTAINS => !str_contains($b, $a),
+            ValueOperator::EQUALS                 => $a === $b,
+            ValueOperator::NOT_EQUALS             => $a !== $b,
+            ValueOperator::GREATER_THAN           => $a > $b,
+            ValueOperator::GREATER_THAN_OR_EQUALS => $a >= $b,
+            ValueOperator::LESS_THAN              => $a < $b,
+            ValueOperator::LESS_THAN_OR_EQUALS    => $a <= $b,
 
             // Unsupported operators
-            ValueOperator::GREATER_THAN, ValueOperator::GREATER_THAN_OR_EQUALS => false,
-            ValueOperator::LESS_THAN, ValueOperator::LESS_THAN_OR_EQUALS => false,
+            ValueOperator::CONTAINS, ValueOperator::NOT_CONTAINS => false,
         };
     }
 }
