+++ src/Glpi/Form/Condition/Type.php
@@ -32,21 +32,11 @@
  * ---------------------------------------------------------------------
  */
 
-namespace Glpi\Form\ConditionalVisiblity;
+namespace Glpi\Form\Condition;
 
-/**
- * Items that implements this interface can be used as a criteria in a condition.
- */
-interface UsedAsCriteriaInterface
+enum Type: string
 {
-    /** @return ValueOperator[] */
-    public function getSupportedValueOperators(): array;
-
-    public function getInputTemplateKey(): InputTemplateKey;
-
-    public function applyValueOperator(
-        mixed $a,
-        ValueOperator $operator,
-        mixed $b,
-    ): bool;
+    case QUESTION = 'question';
+    case SECTION = 'section';
+    case COMMENT = 'comment';
 }
