+++ src/Glpi/Form/Condition/UsedAsCriteriaInterface.php
@@ -32,13 +32,21 @@
  * ---------------------------------------------------------------------
  */
 
-namespace Glpi\Form\ConditionalVisiblity;
+namespace Glpi\Form\Condition;
 
 /**
- * References supported input templates keys in conditional_visibility_editor.html.twig
+ * Items that implements this interface can be used as a criteria in a condition.
  */
-enum InputTemplateKey
+interface UsedAsCriteriaInterface
 {
-    case STRING;
-    case NUMBER;
+    /** @return ValueOperator[] */
+    public function getSupportedValueOperators(): array;
+
+    public function getInputTemplateKey(): InputTemplateKey;
+
+    public function applyValueOperator(
+        mixed $a,
+        ValueOperator $operator,
+        mixed $b,
+    ): bool;
 }
