+++ src/Glpi/Form/Export/Serializer/FormSerializer.php
@@ -39,8 +39,8 @@ use Entity;
 use Glpi\Form\AccessControl\FormAccessControl;
 use Glpi\Form\Category;
 use Glpi\Form\Comment;
-use Glpi\Form\ConditionalVisiblity\ConditionData;
-use Glpi\Form\ConditionalVisiblity\FormData;
+use Glpi\Form\Condition\ConditionData;
+use Glpi\Form\Condition\FormData;
 use Glpi\Form\Destination\FormDestination;
 use Glpi\Form\Export\Context\DatabaseMapper;
 use Glpi\Form\Export\Context\ConfigWithForeignKeysInterface;
