+++ templates/pages/admin/form/form_comment.html.twig
@@ -48,7 +48,7 @@
     role="option"
     data-glpi-form-editor-block
     data-glpi-form-editor-comment
-    data-glpi-form-editor-condition-type="{{ enum('Glpi\\Form\\ConditionalVisiblity\\Type').COMMENT.value }}"
+    data-glpi-form-editor-condition-type="{{ enum('Glpi\\Form\\Condition\\Type').COMMENT.value }}"
     data-glpi-draggable-item
 >
     <section
