+++ templates/pages/admin/form/form_section.html.twig
@@ -50,7 +50,7 @@
 
 <section
     data-glpi-form-editor-section
-    data-glpi-form-editor-condition-type="{{ enum('Glpi\\Form\\ConditionalVisiblity\\Type').SECTION.value }}"
+    data-glpi-form-editor-condition-type="{{ enum('Glpi\\Form\\Condition\\Type').SECTION.value }}"
     class="mt-4"
     aria-label="{{ __("Form section") }}"
 >
